+++ CHANGELOG.md
@@ -19,8 +19,7 @@ IMPROVEMENTS
 * Updating `mark3labs/mcp-go` and `hashicorp/tfe-go` versions. See [#121](https://github.com/hashicorp/terraform-mcp-server/pull/121)
 * Implementing TLS for the http mode of the MCP server. See [#168](https://github.com/hashicorp/terraform-mcp-server/pull/168)
 * Implemented rate limiting with the MCP server. See [#155](https://github.com/hashicorp/terraform-mcp-server/pull/155)
-* Improved and expanded README with usage instructions for Claude Code. See [#173](https://github.com/hashicorp/terraform-mcp-server/pull/173)
-* Improved and expanded README with usage instructions for Cursor. See [#173](https://github.com/hashicorp/terraform-mcp-server/pull/173)
+* Enabled explicit approval for certain tools. See [#172](https://github.com/hashicorp/terraform-mcp-server/pull/172)
 * Improved README with one-click install badges for VSCode/VSCode Insiders/Cursor. See [#173](https://github.com/hashicorp/terraform-mcp-server/pull/173) 
 
 FIXES